+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - add request fairing
 - feat: `func File(filepath string) *Response` at package `response`, use to create a file response with default content-type
 - (#126) fix: duplicate path would panic now
+- (#125) feature: support auto implements OPTIONS method
 
 ## v0.12.9